app/requirements.txt
Django==3.2.6
djangorestframework
django-cors-headers
psycopg2-binary
gunicorn
qrcode
aiohttp
boto3

